Last upload before due date
Worked on getting the tasks to work properly, but was not able to get the leds to toggle. Have added a .txt to the upload as well. Will continue working on this until I get the leds to function properly, and will send in video proof when it is complete. Though it will be late.

+++ Lab2/Lab02/Lab02/src/myButtons.h
@@ -8,4 +8,4 @@
 
 //initializeButtonDriver and readButton
 void initializeButtonDriver(void);
-int readButton(uint8_t uiButtonNum);
+uint8_t readButton(uint8_t uiButtonNum);
